Moved everything expo into client and added package.json and dependencies to client
