+++ client/assets/adaptive-icon.png
