+++ client/package.json
@@ -1,0 +1,31 @@
+{
+  "name": "whimsical-client",
+  "version": "1.0.0",
+  "main": "node_modules/expo/AppEntry.js",
+  "scripts": {
+    "start": "expo start",
+    "android": "expo start --android",
+    "ios": "expo start --ios",
+    "web": "expo start --web",
+    "eject": "expo eject"
+  },
+  "dependencies": {
+    "@react-native-community/masked-view": "^0.1.11",
+    "axios": "^0.24.0",
+    "expo": "~43.0.2",
+    "expo-status-bar": "~1.1.0",
+    "nodemon": "^2.0.15",
+    "react": "17.0.1",
+    "react-dom": "17.0.1",
+    "react-native": "0.64.3",
+    "react-native-gesture-handler": "~1.10.2",
+    "react-native-reanimated": "~2.2.0",
+    "react-native-safe-area-context": "^3.3.2",
+    "react-native-screens": "~3.8.0",
+    "react-native-web": "0.17.1"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.12.9"
+  },
+  "private": false
+}
