+++ package.json
@@ -1,36 +1,28 @@
 {
   "name": "whimsical",
   "version": "1.0.0",
-  "main": "node_modules/expo/AppEntry.js",
+  "main": "index.js",
   "scripts": {
-    "server-dev": "nodemon server/app.js",
-    "start": "expo start",
-    "android": "expo start --android",
-    "ios": "expo start --ios",
-    "web": "expo start --web",
-    "eject": "expo eject"
+    "start": "nodemon server/app.js"
   },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/rpp30-boc-heartbleed/whimsical.git"
+  },
+  "bugs": {
+    "url": "https://github.com/rpp30-boc-heartbleed/whimsical/issues"
+  },
+  "homepage": "https://github.com/rpp30-boc-heartbleed/whimsical#readme",
   "dependencies": {
-    "@react-native-community/masked-view": "^0.1.11",
     "axios": "^0.24.0",
     "dotenv": "^10.0.0",
     "expo": "~43.0.2",
     "expo-status-bar": "~1.1.0",
     "express": "^4.17.1",
     "mongoose": "^6.1.2",
-    "nodemon": "^2.0.15",
-    "react": "17.0.1",
-    "react-dom": "17.0.1",
-    "react-native": "0.64.3",
-    "react-native-gesture-handler": "~1.10.2",
-    "react-native-reanimated": "~2.2.0",
-    "react-native-router-flux": "^4.3.1",
-    "react-native-safe-area-context": "^3.3.2",
-    "react-native-screens": "~3.8.0",
-    "react-native-web": "0.17.1"
+    "nodemon": "^2.0.15"
   },
   "devDependencies": {
-    "@babel/core": "^7.12.9",
     "eslint": "^8.4.1",
     "eslint-config-airbnb": "^19.0.2",
     "eslint-plugin-import": "^2.25.3",
